+++ search/SearchEnvironment.h
@@ -121,7 +121,7 @@ action SearchEnvironment<state,action>::GetAction(const state &s1, const state &
 {
 	std::vector<action> a;
 	GetActions(s1, a);
-	for (int x = 0; x < a.size(); x++)
+	for (size_t x = 0; x < a.size(); x++)
 	{
 		state s = s1;
 		ApplyAction(s, a[x]);
